fake_news_pro/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.7.2
python-multipart==0.0.9
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==4.2.0
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
Pillow==10.3.0
nltk==3.8.1
scikit-learn==1.5.0
numpy==1.26.4
reportlab==4.2.0
aiofiles==23.2.1
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
